Adicionando regras de negocio na Classse Produto Service

+++ src/main/java/br/com/indra/lucas_carlos_batista_cp2026/repository/ProdutosRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProdutosRepository extends JpaRepository<Produtos, Integer> {
+public interface ProdutosRepository extends JpaRepository<Produtos, Long> {
 }
